requirements.txt
# Core dependencies
anthropic>=0.39.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0

# Scraping
outscraper>=3.0.0

# CLI
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0